Refactor fuzzy matching a bit more, make symspell own file and attach original license
